services/transaction-service/app/main.py
from fastapi import FastAPI
from . import routes, models, database
from .database import engine, Base
from dotenv import load_dotenv

load_dotenv()

app = FastAPI()

@app.on_event("startup")
async def startup():
    async with engine.begin() as conn:
        await conn.run_sync(Base.metadata.create_all)

app.include_router(routes.router)
