services/transaction-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
python-dotenv
psycopg2-binary
httpx
aiokafka